# Documentation/MovieReviewApp_API_Production.postman_collection.json-OpenApi3Json.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "MovieReviewApp_API_Production",
    "description": "## Introduction\n\nBackend API for the Glints Academy x Binar Academy mini project batch 14 - D team. Movies review application to manage movies, reviews, users and authentications. Let's do this!",
    "contact": {},
    "version": "1.0"
  },
  "servers": [
    {
      "url": "https://miniproject-moviereviewapp.herokuapp.com",
      "variables": {}
    }
  ],
  "paths": {
...
    "/admin": {
      "get": {
        "tags": [
          "Users (admin only)"
        ],
        "summary": "Get All Users",
        "description": "## Get All Users\nOnly Admin can get all users data from database.",
        "operationId": "GetAllUsers",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "",
            "required": true,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "example": 4
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      },
      "post": {
        "tags": [
          "Users (admin only)"
        ],
        "summary": "Create User",
        "description": "## Create user by admin\nAdmin create an user account to database",
        "operationId": "CreateUser",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserRequest"
              },
              "example": {
                "name": "Udin",
                "email": "udin@gmail.com",
                "password": "123456"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    },
...
    "/admin/61553b8b5c89edea59d05ca2": {
      "put": {
        "tags": [
          "Users (admin only)"
        ],
        "summary": "Update User",
        "description": "## Edit user by admin\nOnly admin can update user account data from database",
        "operationId": "UpdateUser",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRequest"
              },
              "example": {
                "name": "Saprol"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      },
      "delete": {
        "tags": [
          "Users (admin only)"
        ],
        "summary": "Delete User",
        "description": "## Delete user by admin\nOnly admin can delete user account data from database",
        "operationId": "DeleteUser",
        "parameters": [
          {
            "name": "Content-Type",
            "in": "header",
            "description": "JSON Type",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "example": "application/json"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateUserRequest": {
        "title": "CreateUserRequest",
        "required": [
          "name",
          "email",
          "password"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "example": {
          "name": "Udin",
          "email": "udin@gmail.com",
          "password": "123456"
        }
      },
      "UpdateUserRequest": {
        "title": "UpdateUserRequest",
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        },
        "example": {
          "name": "Saprol"
        }
      },
...
    },
    "securitySchemes": {
      "httpBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [
    {
      "httpBearer": []
    }
  ],
...
}